+++ src/strat_backtest/exit_method/half_lifo_exit.py
@@ -36,7 +36,7 @@ class HalfLIFOExit(ExitStruct):
         self,
         open_trades: OpenTrades,
         dt: datetime,
-        exit_price: Decimal,
+        exit_price: float,
     ) -> ClosedPositionResult:
         """Update existing StockTrade objects (still open); and remove completed
         StockTrade objects in 'open_trades'.
@@ -46,7 +46,7 @@ class HalfLIFOExit(ExitStruct):
                 Deque list of StockTrade pydantic object to record open trades.
             dt (datetime):
                 Trade datetime object.
-            exit_price (Decimal):
+            exit_price (float):
                 Exit price of stock ticker.
 
         Returns:
