+++ src/strat_backtest/exit_method/take_all_exit.py
@@ -22,7 +22,7 @@ class TakeAllExit(ExitStruct):
         self,
         open_trades: OpenTrades,
         dt: datetime,
-        exit_price: Decimal,
+        exit_price: float,
     ) -> ClosedPositionResult:
         """Update existing StockTrade objects (still open); and remove completed
         StockTrade objects in 'open_trades'.
@@ -32,7 +32,7 @@ class TakeAllExit(ExitStruct):
                 Deque list of StockTrade pydantic object to record open trades.
             dt (datetime):
                 Trade datetime object.
-            exit_price (Decimal):
+            exit_price (float):
                 Exit price of stock ticker.
 
         Returns:
